Split syntax files into .c and .h files
