+++ modules/syntax/bash/bash.h
@@ -2,25 +2,8 @@
 #define _SYNTAX_BASH_H
 
 // BASH
-char *BASH_extensions[] = {".sh", NULL};
-char *BASH_keywords[] = {
-
-	// Keyword: Misc	
-	"function", "select", "read", "echo", "break", 
-	"export", "let", "clear", "exit",
-	"!", ";;", "|", ">", ">>", "<<", "<", "&",
-
-	// Keyword: Conditionals
-	"if~", "elif~", "else~", "fi~",  "case~", "esac~", "then~",
-
-	// Keyword: Return
-	"return#",
-
-	// Keyword: Loops
-	"for@", "while@", "until@", "in@", "do@", "done@",
-
-	NULL
-};
+extern char *BASH_extensions[];
+extern char *BASH_keywords[];
 
 #define BASH_syntax {	\
 	BASH_extensions,	\
