+++ modules/syntax/brain/brain.c
@@ -1,0 +1,6 @@
+#include <stdlib.h>
+char *BRAIN_extensions[] = {".brain", ".b", NULL};
+char *BRAIN_keywords[] = {
+        "+", "-", "#", ".", ",", ">", "<", "*", "[", "]", "{", "}",
+        "$", "/", "%", "!", "?", ":", ";", "^", NULL
+};
