+++ modules/syntax/ccpp/ccpp.h
@@ -2,27 +2,10 @@
 #define _SYNTAX_CCPP_H
 
 // C and C++
-char *CCPP_extensions[] = {".c", ".cpp", ".h" , ".hpp", NULL};
-char *CCPP_keywords[] = {
-	//types and misc
-	"char", "bool", "short", "int", "__int8", "__int16", "__int32", "__int64",
-	"long", "wchar_t", "__wchar_t", "float", "double", "true", "false",
-	"continue", "break", "enum", "struct","class","default#","namespace",
-	//preprocessor
-	"#define|", "#elif|", "#else|", "#ifndef|", "#error|", "#if|", "#ifdef|",
-	"#pragma|", "#import|", "#include|", "#line|", "#undef|", "#using|", "#endif|",
-	//conditionals
-	"if~", "else~", "switch~", "case~", "try~", "throw~", "catch~",
-	//return
-	"return#", "goto#",
-	//adapters
-	"const^", "static^", "public^", "protected^", "void^", "typedef^", 
-	"union^", "virtual^", "volatile^", "private^", "register^",
-	//loops,
-	"for@", "while@", "do@",
-	NULL
-};
-
+/* Header file should consist of definitions ideally. Including this multiple times is leading to
+ * multiple definition error. */
+extern char *CCPP_extensions[];
+extern char *CCPP_keywords[];
 
 //struct editor_syntax CCPP = {
 //	CCPP_extensions,
@@ -33,7 +16,6 @@ char *CCPP_keywords[] = {
 //	HL_HIGHLIGHT_NUMBERS | HL_HIGHLIGHT_STRINGS
 //};
 
-
 #define CCPP_syntax { \
 	CCPP_extensions, \
 	CCPP_keywords, \
@@ -42,8 +24,5 @@ char *CCPP_keywords[] = {
 	"*/", \
 	HL_HIGHLIGHT_NUMBERS | HL_HIGHLIGHT_STRINGS \
 }
-
-
-
 
 #endif
