+++ modules/syntax/python/python.h
@@ -1,22 +1,8 @@
 #ifndef _SYNTAX_PYTHON_H
 #define _SYNTAX_PYTHON_H
 
-char *Python_extensions[] = {".py", NULL};
-char *Python_keywords[] = {
-	// data types
-	"str", "list", "tuple", "dict", "True", "False", "int", "None", "float",
-	// imports etc.
-	"import|", "from|",
-	//conditionals
-	"if~", "elif~", "continue~", "break~", "pass~", "try~", 
-	// return
-	"return#", "raise#", "is#", "in#",
-	// loops
-	"for@", "while@", "try@", "except@", "finally@",
-	// misc
-	"and^", "or^", "global^", "print^", "with^", "yield^", "not^", "class^", "as^",
-	NULL
-};
+extern char *Python_extensions[];
+extern char *Python_keywords[];
 
 #define Python_syntax { \
 	Python_extensions, \
