+++ modules/syntax/sql/sql.c
@@ -1,0 +1,13 @@
+//
+// Created by supreets51 on 10/12/17.
+//
+#include <stdlib.h>
+// MySQL, just the beginning
+char *SQL_extensions[] = {".sql" ,NULL};
+char *SQL_keywords[] = {
+        "create","insert","from","values","select","alter","orderby",
+        "table~","database~","column~",
+        "asc@","desc@",
+        NULL
+};
+
